backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
autopep8==2.0.0
cffi==1.15.1
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.1.5.post3
MarkupSafe==2.1.1
packaging==21.3
protobuf==3.20.1
pycodestyle==2.10.0
pycparser==2.21
pyparsing==3.0.9
tomli==2.0.1
Werkzeug==2.2.2
